[vtbgen] fix issue with intger CYCLE_TIME
We were using this in our generated Verilog test bench:

  always #(`CYCLE_TIME/2) clk = ~clk;

But if CYCLE_TIME is an odd integer (say 9) this would result in integer
division and round down (i.e., with a cycle time of 9 we would end up with
a clock period of 8!). This doesn't matter unless we are doing back-
annotated gate-level simulation _and_ we use an odd integer clock period
which is why we never detected this bug before. The fix is to simply
force floating point division.

  always #((`CYCLE_TIME*1.0)/2) clk = ~clk;