Fix issue with changing vcd file name multiple times
We were seeing a bug where if you ran a bunch of tests with --dump-vcd
PyMTL3 would only dump the VCD for the first test, although every once
and a while it would indeed dump VCD files for the every test. This
seems to have something to do with the shared library. The cleanest
fix seems to be to avoid hard coding the VCD file name in the generated
PyMTL wrapper and instead to dynamically get the VCD file name from
the metadata.