Fix Verilator segfaults due to multi-instantiation
We had to do this kind of craziness to avoid Verilator segfault:

  try:

    t( 0x00, '?'  )
    t( 0x13, 0x01 )
    t( 0x27, 0x14 )
    t( 0x00, 0x28 )
    t( 0x00, 0x01 )
    t( 0x00, 0x01 )

  finally:

    model.finalize()

And we saw segfaults when we tried instantiating the same Verilog
module twice and then using PyMTL for the composition. Peitian
dug into this and provided a fix.

Peitian writes: "This is a change to the C wrapper such that we don't
always create a new VerilatedContext everytime we need to create a new
instance of the model; instead, this change uses a globally shared
VerilatedContext guarded by a reference counter. Based on what I saw I
think we were not managing contexts correctly... This specific segfault
only seems to happen when two instances of the same model are finalized
back to back."